+++ internal/httpserve/handlers/ent.go
@@ -6,7 +6,6 @@ import (
 
 	gowebauthn "github.com/go-webauthn/webauthn/webauthn"
 	"github.com/rs/zerolog/log"
-	"github.com/stripe/stripe-go/v81"
 
 	ent "github.com/theopenlane/core/internal/ent/generated"
 	"github.com/theopenlane/core/internal/ent/generated/emailverificationtoken"
@@ -526,50 +525,6 @@ func (h *Handler) getOrgSettingByOrgID(ctx context.Context, orgID string) (*ent.
 	}
 
 	return setting, nil
-}
-
-func (h *Handler) fetchOrCreateStripe(context context.Context, orgsetting *ent.OrganizationSetting) (*stripe.Customer, error) {
-	if orgsetting.BillingEmail == "" {
-		log.Error().Msgf("billing email is required to be set to create a checkout session")
-		return nil, ErrNoBillingEmail
-	}
-
-	if orgsetting.StripeID != "" {
-		cust, err := h.Entitlements.Client.Customers.Get(orgsetting.StripeID, nil)
-		if err != nil {
-			log.Error().Err(err).Msg("error fetching stripe customer")
-			return nil, err
-		}
-
-		if cust.Email != orgsetting.BillingEmail {
-			log.Error().Msgf("customer email does not match, updating stripe customer")
-
-			_, err := h.Entitlements.Client.Customers.Update(orgsetting.StripeID, &stripe.CustomerParams{
-				Email: &orgsetting.BillingEmail,
-			})
-			if err != nil {
-				log.Error().Err(err).Msg("error updating stripe customer")
-				return nil, err
-			}
-		}
-
-		return cust, nil
-	}
-
-	stripeCustomer, err := h.Entitlements.Client.Customers.New(&stripe.CustomerParams{
-		Email: &orgsetting.BillingEmail,
-	})
-	if err != nil {
-		log.Error().Err(err).Msg("error creating stripe customer")
-		return nil, err
-	}
-
-	if err := h.updateOrganizationSettingWithCustomerID(context, orgsetting.ID, stripeCustomer.ID); err != nil {
-		log.Error().Err(err).Msg("error updating organization setting with stripe customer id")
-		return nil, err
-	}
-
-	return stripeCustomer, nil
 }
 
 func (h *Handler) updateOrganizationSettingWithCustomerID(ctx context.Context, orgsettingID, customerID string) error {
