+++ pkg/events/soiree/options.go
@@ -60,6 +60,7 @@ func WithErrChanBufferSize(size int) EventPoolOption {
 	}
 }
 
+// WithClient sets a custom client for the Soiree
 func WithClient(client interface{}) EventPoolOption {
 	return func(m Soiree) {
 		m.SetClient(client)
